+++ src/services/helpers/renderService.js
@@ -1,5 +1,5 @@
 import StyleService from '../../services/dom/styleService';
-import { IdService } from '../dom/idService';
+import { IdService, LoggerService } from '../../services';
 
 export class RenderService {
     static showComponent(isEnabled, template) {
@@ -7,6 +7,17 @@ export class RenderService {
             return template;
         }
         return '';
+    }
+
+    static modal(el, tm = 500) {
+        setTimeout(() => {
+            if (el) {
+              el.close();
+              el.showModal();
+            } else {
+              LoggerService.warn(`Failed to open #${el} -> modal lost focus`);
+            }
+        }, tm);
     }
 
     static toggleButton(id, context, timeout = 2000) {