+++ Changelog.md
@@ -1,4 +1,7 @@
 ## Changelog
+v 2.51
+  Added improvement to fix lost focus of showModal
+
 v 2.50
   Improved Services utils to use classes, Added removeEvent and article paragraph
    not element in Node issue fixes, Added showModal issue fixes